refactor(server): pull integration specificatio...
...n handler into separate class
