+++ app/server/endpoint/src/main/java/io/syndesis/server/endpoint/v1/handler/integration/IntegrationHandler.java
@@ -15,20 +15,17 @@
  */
 package io.syndesis.server.endpoint.v1.handler.integration;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
 import javax.persistence.EntityNotFoundException;
 import javax.validation.Validator;
+import javax.validation.constraints.NotNull;
 import javax.validation.groups.ConvertGroup;
 import javax.validation.groups.Default;
-import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -39,28 +36,22 @@ import javax.ws.rs.core.UriInfo;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
 import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
+import io.swagger.annotations.ApiParam;
 import io.syndesis.common.model.DataShape;
 import io.syndesis.common.model.Kind;
 import io.syndesis.common.model.ListResult;
 import io.syndesis.common.model.filter.FilterOptions;
 import io.syndesis.common.model.filter.Op;
-import io.syndesis.common.model.integration.Flow;
 import io.syndesis.common.model.integration.Integration;
 import io.syndesis.common.model.integration.IntegrationDeployment;
 import io.syndesis.common.model.integration.IntegrationOverview;
-import io.syndesis.common.model.integration.Step;
-import io.syndesis.common.model.openapi.OpenApi;
 import io.syndesis.common.model.validation.AllValidations;
 import io.syndesis.server.api.generator.APIGenerator;
-import io.syndesis.server.api.generator.APIIntegration;
 import io.syndesis.server.dao.manager.DataManager;
 import io.syndesis.server.dao.manager.EncryptionComponent;
 import io.syndesis.server.endpoint.util.PaginationFilter;
 import io.syndesis.server.endpoint.util.ReflectiveSorter;
 import io.syndesis.server.endpoint.v1.handler.BaseHandler;
-import io.syndesis.server.endpoint.v1.handler.api.ApiGeneratorHelper;
-import io.syndesis.server.endpoint.v1.handler.api.ApiHandler;
 import io.syndesis.server.endpoint.v1.operations.Creator;
 import io.syndesis.server.endpoint.v1.operations.Deleter;
 import io.syndesis.server.endpoint.v1.operations.Getter;
@@ -72,7 +63,6 @@ import io.syndesis.server.endpoint.v1.operations.Validating;
 import io.syndesis.server.inspector.Inspectors;
 import io.syndesis.server.openshift.OpenShiftService;
 
-import org.jboss.resteasy.annotations.providers.multipart.MultipartForm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
@@ -87,7 +77,7 @@ public class IntegrationHandler extends BaseHandler implements Lister<Integratio
 
     private static final Logger LOGGER = LoggerFactory.getLogger(IntegrationHandler.class);
 
-    private final APIGenerator apiGenerator;
+    final APIGenerator apiGenerator;
     private final EncryptionComponent encryptionSupport;
     private final Inspectors inspectors;
     private final OpenShiftService openShiftService;
@@ -239,100 +229,9 @@ public class IntegrationHandler extends BaseHandler implements Lister<Integratio
         getDataManager().update(updatedIntegration);
     }
 
-    @PUT
-    @Consumes(MediaType.MULTIPART_FORM_DATA)
-    @Produces(MediaType.APPLICATION_JSON)
     @Path("/{id}/specification")
-    @ApiOperation("For an integration that is generated from a specification updates it so it conforms to the updated specification")
-    public void updateSpecification(@PathParam("id") final String id, @MultipartForm final ApiHandler.APIFormData apiFormData) {
-        final Integration existing = getIntegration(id);
-
-        final APIIntegration apiIntegration = ApiGeneratorHelper.generateIntegrationUpdateFrom(existing, apiFormData, getDataManager(), apiGenerator);
-
-        final Integration givenIntegration = apiIntegration.getIntegration();
-
-
-        final Integration updated = updateFlowsAndStartAndEndDataShapes(existing, givenIntegration);
-
-        if (existing.getFlows().equals(updated.getFlows())) {
-            // no changes were made to the flows
-            return;
-        }
-
-        // store the OpenAPI resource, we keep the old one as it might
-        // be referenced from Integration's stored in IntegrationDeployent's
-        // this gives us a rollback mechanism
-        getDataManager().store(apiIntegration.getSpec(), OpenApi.class);
-
-        // perform the regular update
-        update(id, updated);
+    public IntegrationSpecificationHandler specification(@NotNull @PathParam("id") @ApiParam(required = true, example = "integration-id") final String id) {
+        return new IntegrationSpecificationHandler(this);
     }
 
-    static Integration updateFlowsAndStartAndEndDataShapes(final Integration existing, final Integration given) {
-        // will contain updated flows
-        final List<Flow> updatedFlows = new ArrayList<>(given.getFlows().size());
-
-        for (final Flow givenFlow : given.getFlows()) {
-            final String flowId = givenFlow.getId().get();
-
-            final Optional<Flow> maybeExistingFlow = existing.findFlowById(flowId);
-            if (!maybeExistingFlow.isPresent()) {
-                // this is a flow generated from a new operation or it
-                // has it's operation id changed, either way we only
-                // need to add it, since we don't know what flow we need
-                // to update
-                updatedFlows.add(givenFlow);
-                continue;
-            }
-
-
-            final List<Step> givenSteps = givenFlow.getSteps();
-            if (givenSteps.size() != 2) {
-                throw new IllegalArgumentException("Expecting to get exactly two steps per flow");
-            }
-
-            // this is a freshly minted flow from the specification
-            // there should be only two steps (start and end) in the
-            // flow
-            final Step givenStart = givenSteps.get(0);
-            final Optional<DataShape> givenStartDataShape = givenStart.outputDataShape();
-
-            // generated flow has only a start and an end, start is at 0
-            // and the end is at 1
-            final Step givenEnd = givenSteps.get(1);
-            final Optional<DataShape> givenEndDataShape = givenEnd.inputDataShape();
-
-            final Flow existingFlow = maybeExistingFlow.get();
-            final List<Step> existingSteps = existingFlow.getSteps();
-
-            // readability
-            final int start = 0;
-            final int end = existingSteps.size() - 1;
-
-            // now we update the data shapes of the start and end steps
-            final Step existingStart = existingSteps.get(start);
-            final Step updatedStart = existingStart.updateOutputDataShape(givenStartDataShape);
-
-            final Step existingEnd = existingSteps.get(end);
-            final Step updatedEnd = existingEnd.updateInputDataShape(givenEndDataShape);
-
-            final List<Step> updatedSteps = new ArrayList<>(existingSteps);
-            updatedSteps.set(start, updatedStart);
-            updatedSteps.set(end, updatedEnd);
-
-            final Flow updatedFlow = existingFlow.builder()
-                .name(givenFlow.getName())
-                .description(givenFlow.getDescription())
-                .steps(updatedSteps)
-                .build();
-            updatedFlows.add(updatedFlow);
-        }
-
-        return existing.builder()
-            .flows(updatedFlows)
-            // we replace all resources counting that the only resource
-            // present is the OpenAPI specification
-            .resources(given.getResources())
-            .build();
-    }
 }
