+++ app/server/endpoint/src/test/java/io/syndesis/server/endpoint/v1/handler/integration/IntegrationHandlerTest.java
@@ -15,7 +15,6 @@
  */
 package io.syndesis.server.endpoint.v1.handler.integration;
 
-import java.io.ByteArrayInputStream;
 import java.security.Principal;
 import java.util.Arrays;
 import java.util.Collections;
@@ -27,30 +26,19 @@ import javax.ws.rs.core.SecurityContext;
 
 import io.syndesis.common.model.DataShape;
 import io.syndesis.common.model.DataShapeKinds;
-import io.syndesis.common.model.action.ConnectorAction;
-import io.syndesis.common.model.action.ConnectorDescriptor;
-import io.syndesis.common.model.connection.Connection;
-import io.syndesis.common.model.connection.Connector;
 import io.syndesis.common.model.filter.FilterOptions;
-import io.syndesis.common.model.integration.Flow;
 import io.syndesis.common.model.integration.Integration;
 import io.syndesis.common.model.integration.IntegrationDeployment;
 import io.syndesis.common.model.integration.IntegrationDeploymentState;
-import io.syndesis.common.model.integration.Step;
-import io.syndesis.common.model.openapi.OpenApi;
 import io.syndesis.server.api.generator.APIGenerator;
-import io.syndesis.server.api.generator.APIIntegration;
-import io.syndesis.server.api.generator.ProvidedApiTemplate;
 import io.syndesis.server.dao.manager.DataManager;
 import io.syndesis.server.dao.manager.EncryptionComponent;
-import io.syndesis.server.endpoint.v1.handler.api.ApiHandler.APIFormData;
 import io.syndesis.server.inspector.Inspectors;
 import io.syndesis.server.openshift.OpenShiftService;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
-import org.mockito.ArgumentMatchers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -101,44 +89,6 @@ public class IntegrationHandlerTest {
     }
 
     @Test
-    public void shouldAddNewFlowsNonTrivialCase() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow flow1 = new Flow.Builder().id("flow1").addStep(step, step).build();
-        final Flow flow2 = new Flow.Builder().id("flow2").addStep(step, step).build();
-        final Flow flow3 = new Flow.Builder().id("flow3").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().addFlow(flow1, flow3).build();
-        final Integration given = new Integration.Builder().addFlow(flow1, flow2, flow3).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated).as("there should be three flows").isEqualTo(given);
-    }
-
-    @Test
-    public void shouldAddNewFlowsTrivialCase() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow flow = new Flow.Builder().id("flow1").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().build();
-        final Integration given = new Integration.Builder().addFlow(flow).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated).as("there should be one flow").isEqualTo(given);
-    }
-
-    @Test
     public void shouldCreateIntegrations() {
         final SecurityContext security = mock(SecurityContext.class);
         final Principal principal = mock(Principal.class);
@@ -158,43 +108,6 @@ public class IntegrationHandlerTest {
         assertThat(persisted.getValue()).isEqualToIgnoringGivenFields(encrypted, "createdAt");
 
         verify(encryptionSupport).encrypt(same(integration));
-    }
-
-    @Test
-    public void shouldDeleteFlowsThatHaveBeenRemovedNonTrivialCase() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow existingFlow1 = new Flow.Builder().id("flow1").addStep(step, step).build();
-        final Flow existingFlow2 = new Flow.Builder().id("flow2").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().addFlow(existingFlow1, existingFlow2).build();
-        final Integration given = new Integration.Builder().addFlow(existingFlow2).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated).as("there should be only one flow").isEqualTo(given);
-    }
-
-    @Test
-    public void shouldDeleteFlowsThatHaveBeenRemovedTrivialCase() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow existingFlow = new Flow.Builder().id("flow1").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().addFlow(existingFlow).build();
-        final Integration given = new Integration.Builder().build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated.getFlows()).as("there shouldn't be any flows as we removed the single existing flow").isEmpty();
     }
 
     @Test
@@ -232,150 +145,6 @@ public class IntegrationHandlerTest {
             .build().deleted().deleted());
         verify(openShiftService).delete("first to delete");
         verify(openShiftService).delete("second to delete");
-    }
-
-    @Test
-    public void shouldPerformUpdatesBasedOnNewSpecification() {
-        final Integration existing = new Integration.Builder().id("integration-1").addFlow(new Flow.Builder().id("integration-1:flows:flow1").build()).build();
-        final Integration given = new Integration.Builder().id("integration-2").addFlow(new Flow.Builder().id("integration-2:flows:flow2").build()).build();
-        final Integration expected = new Integration.Builder().id("integration-1").addFlow(new Flow.Builder().id("integration-1:flows:flow2").build()).build();
-
-        final OpenApi updatedSpecification = new OpenApi.Builder().build();
-        final APIIntegration updatedApiIntegration = new APIIntegration(given, updatedSpecification);
-
-        when(dataManager.fetch(Connection.class, "api-provider")).thenReturn(new Connection.Builder().connectorId("api-provider-connector").build());
-        when(dataManager.fetch(Connector.class, "api-provider-connector")).thenReturn(new Connector.Builder().build());
-        when(dataManager.fetch(Integration.class, "integration-1")).thenReturn(existing);
-        when(encryptionSupport.encrypt(expected)).thenReturn(expected);
-        when(apiGenerator.generateIntegration(any(String.class), any(ProvidedApiTemplate.class))).thenReturn(updatedApiIntegration);
-
-        final APIFormData openApiUpdate = new APIFormData();
-        openApiUpdate.setSpecification(new ByteArrayInputStream("updated specification".getBytes()));
-
-        handler.updateSpecification("integration-1", openApiUpdate);
-
-        verify(dataManager).store(updatedSpecification, OpenApi.class);
-        verify(dataManager).update(ArgumentMatchers.<Integration> argThat(v -> {
-            assertThat(v).isEqualToIgnoringGivenFields(expected, "version", "updatedAt");
-            assertThat(v.getVersion()).isEqualTo(2);
-            return true;
-        }));
-    }
-
-    @Test
-    public void shouldUpdateFlowNameAndDescription() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow flow = new Flow.Builder().id("flow1").name("name").description("description").addStep(step, step).build();
-        final Flow flowUpdated = new Flow.Builder().id("flow1").name("updated name").description("updated description").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().addFlow(flow).build();
-        final Integration given = new Integration.Builder().addFlow(flowUpdated).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated).as("name and description should be updated").isEqualTo(given);
-    }
-
-    @Test
-    public void shouldUpdateFlowsDataShapes() {
-        final Step startStep = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder()
-                    .outputDataShape(new DataShape.Builder()
-                        .kind(DataShapeKinds.JSON_INSTANCE)
-                        .specification("{\"start\":\"existing\"}")
-                        .build())
-                    .build())
-                .build())
-            .build();
-
-        final Step endStep = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder()
-                    .inputDataShape(new DataShape.Builder()
-                        .kind(DataShapeKinds.JSON_INSTANCE)
-                        .specification("{\"end\":\"existing\"}")
-                        .build())
-                    .build())
-                .build())
-            .build();
-
-        final Step stepWithShapes = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder()
-                    .inputDataShape(new DataShape.Builder()
-                        .kind(DataShapeKinds.ANY)
-                        .build())
-                    .outputDataShape(new DataShape.Builder()
-                        .kind(DataShapeKinds.NONE)
-                        .build())
-                    .build())
-                .build())
-            .build();
-
-
-        final Flow existingFlow = new Flow.Builder().id("flow1").addStep(startStep, stepWithShapes, endStep).build();
-        final Integration existing = new Integration.Builder().addFlow(existingFlow).build();
-
-        final DataShape givenStartShape = new DataShape.Builder()
-            .kind(DataShapeKinds.JSON_INSTANCE)
-            .specification("{\"start\":\"updated\"}")
-            .build();
-
-        final DataShape givenEndShape = new DataShape.Builder()
-            .kind(DataShapeKinds.JSON_INSTANCE)
-            .specification("{\"end\":\"updated\"}")
-            .build();
-
-        final Step givenStartStep = startStep.updateOutputDataShape(Optional.of(givenStartShape));
-        final Step givenEndStep = endStep.updateInputDataShape(Optional.of(givenEndShape));
-        final Flow givenFlow = new Flow.Builder()
-            .id("flow1")
-            .steps(Arrays.asList(
-                givenStartStep,
-                givenEndStep
-            ))
-            .build();
-        final Integration given = new Integration.Builder().addFlow(givenFlow).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        final Flow expectedFlow = new Flow.Builder()
-            .id("flow1")
-            .steps(Arrays.asList(
-                givenStartStep,
-                stepWithShapes,
-                givenEndStep
-            ))
-            .build();
-        final Integration expected = existing.builder()
-            .flows(Collections.singleton(expectedFlow))
-            .build();
-
-        assertThat(updated).as("should update only the data shapes").isEqualTo(expected);
-    }
-
-    @Test
-    public void shouldUpdateFlowsStartAndEndDataShapesWithoutChanges() {
-        final Step step = new Step.Builder()
-            .action(new ConnectorAction.Builder()
-                .descriptor(new ConnectorDescriptor.Builder().build())
-                .build())
-            .build();
-
-        final Flow flow = new Flow.Builder().id("flow1").addStep(step, step).build();
-
-        final Integration existing = new Integration.Builder().addFlow(flow).build();
-        final Integration given = new Integration.Builder().addFlow(flow).build();
-
-        final Integration updated = IntegrationHandler.updateFlowsAndStartAndEndDataShapes(existing, given);
-
-        assertThat(updated).as("there should be no changes in trivial case").isEqualTo(existing);
     }
 
     private static DataShape dataShape(DataShapeKinds kind) {
